requirements.txt
discord
boto3
load_dotenv